Updating setup instructions to use NuGet Package
Removing references to old way of referencing the Windows SDK, and added new instructions on installing the NuGet package. This eliminated a step.  

Also added instructions to install the new desktop notifications community toolkit. I also removed references to the copy/paste instructions.

We are still waiting for the link to the notifications package, that will be updated when this has been reviewed.